use configuration static directly
This gets rid of te config variable in all the structs by using the
static variable in the basic template. To be able to access the variable
more easily use a once_cell::sync::Lazy instead of a OnceCEll because
Lazy implements Deref letting us directly access the struct fields.