use Solarized colour scheme, dark mode
To be able to use dark mode in the form of CSS's prefers-color-scheme,
syntect had to be switched to output classes instead of hard coded colours.

This is a bit problematic as seen in the changes on the stylesheet. Some of
the CSS classes could probably be merged, but this is a nontrivial problem.

Because of the increased size, we should enable Caching for the stylesheet,
which tide's serve_file does not automatically do.