fix CSS for syntax highlighting
I forgot to change the class style back to Spaced, which is what
was used when generating the base of the style sheets from syntect.

Also the .line CSS selector interfered with some syntect items
that were also using this class. The solution is to restrict it
to links only. This made some lines, e.g. with doc comments
display wrong by having multiple line numbers on one line.