Allow selection of only repo url
Current behavior only allows for the selection of the full text, with an added <a> tag it selects all for the first click, double click selects only the URL.

On chromium it doesn't seem to affect anything outside of the css colors. Double click behavior was tested with Firefox and chromium on linux.